+++ firebase-database/src/main/java/com/androidhuman/rxfirebase2/database/ChildEvent.java
@@ -2,7 +2,7 @@ package com.androidhuman.rxfirebase2.database;
 
 import com.google.firebase.database.DataSnapshot;
 
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 
 public abstract class ChildEvent {