+++ firebase-database/src/main/java/com/androidhuman/rxfirebase2/database/ChildMoveEvent.java
@@ -3,9 +3,9 @@ package com.androidhuman.rxfirebase2.database;
 import com.google.auto.value.AutoValue;
 import com.google.firebase.database.DataSnapshot;
 
-import android.support.annotation.CheckResult;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
+import androidx.annotation.CheckResult;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 
 @AutoValue
 public abstract class ChildMoveEvent extends ChildEvent {