+++ firebase-database/src/main/java/com/androidhuman/rxfirebase2/database/ChildRemoveEvent.java
@@ -3,8 +3,8 @@ package com.androidhuman.rxfirebase2.database;
 import com.google.auto.value.AutoValue;
 import com.google.firebase.database.DataSnapshot;
 
-import android.support.annotation.CheckResult;
-import android.support.annotation.NonNull;
+import androidx.annotation.CheckResult;
+import androidx.annotation.NonNull;
 
 @AutoValue
 public abstract class ChildRemoveEvent extends ChildEvent {